Added buffered_collection
